+++ crates/inspector/Cargo.toml
