+++ crates/primitives/src/lib.rs
@@ -7,33 +7,16 @@
 #[cfg(not(feature = "std"))]
 extern crate alloc as std;
 
-pub mod block;
-pub mod db;
-pub mod eip7702;
-pub mod env;
-
 mod constants;
-mod evm_wiring;
 #[cfg(any(feature = "c-kzg", feature = "kzg-rs"))]
 pub mod kzg;
-pub mod precompile;
-pub mod result;
 pub mod specification;
-pub mod transaction;
-pub mod utilities;
-pub use alloy_eips::eip2930::{AccessList, AccessListItem};
 pub use alloy_primitives::{
-    self, address, b256, bytes, fixed_bytes, hex, hex_literal, ruint, uint, Address, Bytes,
-    FixedBytes, Log, LogData, TxKind, B256, I256, U256,
+    self, address, b256, bytes, fixed_bytes, hex, hex_literal, keccak256, ruint, uint, Address,
+    Bytes, FixedBytes, Log, LogData, TxKind, B256, I256, U256,
 };
 pub use bitvec;
 pub use constants::*;
-pub use eip7702::{
-    Authorization, AuthorizationList, InvalidAuthorization, RecoveredAuthorization, Signature,
-    SignedAuthorization,
-};
-pub use env::*;
-pub use evm_wiring::*;
 
 cfg_if::cfg_if! {
     if #[cfg(all(not(feature = "hashbrown"), feature = "std"))] {
@@ -44,15 +27,14 @@ cfg_if::cfg_if! {
     }
 }
 
-pub use block::Block;
 #[cfg(any(feature = "c-kzg", feature = "kzg-rs"))]
 pub use kzg::{EnvKzgSettings, KzgSettings};
-pub use precompile::*;
-pub use result::*;
 pub use specification::*;
-pub use transaction::Transaction;
-pub use utilities::*;
 
 #[cfg(all(feature = "c-kzg", feature = "kzg-rs"))]
 // silence kzg-rs lint as c-kzg will be used as default if both are enabled.
 use kzg_rs as _;
+
+/// The Keccak-256 hash of the empty string `""`.
+pub const KECCAK_EMPTY: B256 =
+    b256!("c5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470");
