+++ crates/specification/src/eip170.rs
@@ -1,0 +1,6 @@
+//! EIP-170: Contract code size limit
+
+/// EIP-170: Contract code size limit
+///
+/// By default the limit is `0x6000` (~25kb)
+pub const MAX_CODE_SIZE: usize = 0x6000;
