+++ crates/specification/src/eip2930.rs
@@ -1,0 +1,1 @@
+pub use alloy_eips::eip2930::{AccessList, AccessListItem};
