+++ crates/specification/src/eip7702.rs
@@ -1,0 +1,10 @@
+mod authorization_list;
+mod constants;
+mod recovered_authorization;
+
+pub use authorization_list::*;
+pub use constants::*;
+pub use recovered_authorization::*;
+
+
+pub use alloy_eips::eip7702::{Authorization, SignedAuthorization};
