+++ crates/specification/src/eip7702/authorization_list.rs
@@ -1,9 +1,7 @@
-pub use alloy_eips::eip7702::{Authorization, SignedAuthorization};
+use super::RecoveredAuthorization;
+use crate::{eip2::SECP256K1N_HALF, eip7702::SignedAuthorization};
 pub use alloy_primitives::{Parity, Signature};
-
-use super::SECP256K1N_HALF;
-use crate::Address;
-use core::{fmt, ops::Deref};
+use core::fmt;
 use std::{boxed::Box, vec::Vec};
 
 /// Authorization list for EIP-7702 transaction type.
@@ -62,7 +60,7 @@ impl AuthorizationList {
             Self::Signed(signed) => signed.iter().try_for_each(validate)?,
             Self::Recovered(recovered) => recovered
                 .iter()
-                .map(|recovered| &recovered.inner)
+                .map(|recovered| recovered.inner())
                 .try_for_each(validate)?,
         };
 
@@ -93,49 +91,6 @@ impl AuthorizationList {
             return self;
         };
         Self::Recovered(signed.into_iter().map(|signed| signed.into()).collect())
-    }
-}
-
-/// A recovered authorization.
-#[derive(Debug, Clone, Hash, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-pub struct RecoveredAuthorization {
-    #[cfg_attr(feature = "serde", serde(flatten))]
-    inner: SignedAuthorization,
-    authority: Option<Address>,
-}
-
-impl RecoveredAuthorization {
-    /// Instantiate without performing recovery. This should be used carefully.
-    pub const fn new_unchecked(inner: SignedAuthorization, authority: Option<Address>) -> Self {
-        Self { inner, authority }
-    }
-
-    /// Get the `authority` for the authorization.
-    ///
-    /// If this is `None`, then the authority could not be recovered.
-    pub const fn authority(&self) -> Option<Address> {
-        self.authority
-    }
-
-    /// Splits the authorization into parts.
-    pub const fn into_parts(self) -> (SignedAuthorization, Option<Address>) {
-        (self.inner, self.authority)
-    }
-}
-
-impl From<SignedAuthorization> for RecoveredAuthorization {
-    fn from(signed_auth: SignedAuthorization) -> Self {
-        let authority = signed_auth.recover_authority().ok();
-        Self::new_unchecked(signed_auth, authority)
-    }
-}
-
-impl Deref for RecoveredAuthorization {
-    type Target = Authorization;
-
-    fn deref(&self) -> &Self::Target {
-        &self.inner
     }
 }
