+++ crates/specification/src/eip7702/constants.rs
@@ -1,0 +1,7 @@
+//! EIP-7702 constants.
+
+// Base cost of updating authorized account.
+pub const PER_AUTH_BASE_COST: u64 = 2500;
+
+/// Cost of creating authorized account that was previously empty.
+pub const PER_EMPTY_ACCOUNT_COST: u64 = 25000;
