+++ crates/specification/src/lib.rs
@@ -1,0 +1,12 @@
+//! Optimism-specific constants, types, and helpers.
+#![cfg_attr(not(test), warn(unused_crate_dependencies))]
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(not(feature = "std"))]
+extern crate alloc as std;
+
+pub mod eip170;
+pub mod eip2;
+pub mod eip2930;
+pub mod eip7702;
+pub mod hardfork;
