+++ crates/transaction/Cargo.toml
@@ -4,7 +4,7 @@ description = "Revm Database interface"
 edition = "2021"
 keywords = ["ethereum", "evm", "revm", "no_std"]
 license = "MIT"
-name = "revm-database-traits"
+name = "revm-transaction"
 repository = "https://github.com/bluealloy/revm"
 version = "1.0.0"
 readme = "../../README.md"
