+++ crates/wiring/src/result.rs
